feat: extract yaml functions
